Address review: dedup runner, normalize types, fix matmul out_dtype
- runner.py: drop redundant _guarded docstring, collapse _rms_column dead
  branch, drop unused seed param, unify duplicated memory-measurement
  blocks, document CPU fallback and L2-flush stream ordering.
- bench_*.py: replace `inputs: dict` with `inputs: Inputs` for consistency.
- bench_rotary: route eager/compiled/compiled_max through
  standard_pytorch_variants instead of building four Variants by hand.
- utils.py: tighten make_grad_reset docstring on grad_buffer + flag.
- sparse_linear.py: add out_dtype=tl.float32 to the last tl.dot call in
  dense_matmul_kernel for symmetry with the other inner-loop dots.

+++ fast_llm/functional/triton/sparse_linear.py
@@ -131,7 +131,7 @@ def dense_matmul_kernel(
                 out_dtype=tl.float32,
             )
         else:
-            out += tl.dot(tl.load(lhs_ptr), tl.load(rhs_ptr))
+            out += tl.dot(tl.load(lhs_ptr), tl.load(rhs_ptr), out_dtype=tl.float32)
 
     # Output
     if masked:
